Logging, Pagination, Elasticsearch improvements

+++ Application/Services/EmployeeService.cs
@@ -3,21 +3,32 @@ using Application.DTOs.Response;
 using Application.Interfaces;
 using Application.Interfaces.Repositories;
 using Domain.Entities;
+using Microsoft.Extensions.Logging;
 
 namespace Application.Services;
 
 public class EmployeeService : IEmployeeService
 {
     private readonly IEmployeeRepository _repository;
+    private readonly ILogger<EmployeeService> _logger;
 
-    public EmployeeService(IEmployeeRepository repository)
+    public EmployeeService(IEmployeeRepository repository, ILogger<EmployeeService> logger)
     {
         _repository = repository;
+        _logger = logger;
     }
 
-    public async Task<List<EmployeeResponse>> GetAllAsync()
+    public async Task<List<EmployeeResponse>> GetAllAsync(int pageNumber, int pageSize)
     {
-        var employees = await _repository.GetAllAsync();
+        _logger.LogInformation("Fetching categories. Page: {PageNumber}, Size: {PageSize}", pageNumber, pageSize);
+        
+        if (pageNumber <= 0 || pageSize <= 0)
+        {
+            _logger.LogWarning("Invalid pagination parameters. Page: {PageNumber}, Size: {PageSize}", pageNumber, pageSize);
+            return new List<EmployeeResponse>();
+        }
+        
+        var employees = await _repository.GetAllAsync(pageNumber, pageSize);
 
         return employees.Select(e => new EmployeeResponse
         {
@@ -35,6 +46,7 @@ public class EmployeeService : IEmployeeService
 
         if (employee == null)
         {
+            _logger.LogWarning("Employee with ID {EmployeeId} not found", id);
             return null;
         }
 
@@ -50,6 +62,9 @@ public class EmployeeService : IEmployeeService
 
     public async Task<EmployeeResponse> CreateAsync(CreateEmployeeRequest request)
     {
+        _logger.LogInformation("Creating employee. Name: {EmployeeName}, Surname: {EmployeeSurname}", request.FirstName,
+            request.LastName);
+        
         var employee = new Employee
         {
             FirstName = request.FirstName,
@@ -59,6 +74,8 @@ public class EmployeeService : IEmployeeService
         };
 
         var created = await _repository.AddAsync(employee);
+        
+        _logger.LogInformation("Employee created successfully with ID: {EmployeeId}", created.Id);
 
         return new EmployeeResponse
         {
@@ -72,20 +89,31 @@ public class EmployeeService : IEmployeeService
 
     public async Task<EmployeeResponse?> UpdateAsync(int id, UpdateEmployeeRequest request)
     {
-        var employee = new Employee
-        {
-            Id = id,
-            FirstName = request.FirstName,
-            LastName = request.LastName,
-            Position = request.Position
-        };
+        _logger.LogInformation("Updating employee with ID: {EmployeeId}", id);
+        
+        var employee = await _repository.GetByIdAsync(id);
 
+        if (employee == null)
+        {
+            _logger.LogWarning("Employee with ID {CustomerId} not found", id);
+            return null;
+        }
+
+        employee.Id = id;
+        employee.FirstName = request.FirstName;
+        employee.LastName = request.LastName;
+        employee.Position = request.Position;
+        
         var updated = await _repository.UpdateAsync(employee);
 
         if (updated == null)
         {
             return null;
         }
+        
+        _logger.LogInformation(
+            "Employee {EmployeeId} updated successfully. Name: {EmployeeName}, Surname: {EmployeeSurname}", id,
+            updated.FirstName, updated.LastName);
 
         return new EmployeeResponse
         {
@@ -97,8 +125,28 @@ public class EmployeeService : IEmployeeService
         };
     }
 
-    public async Task DeleteAsync(int id)
+    public async Task<EmployeeResponse?> DeleteAsync(int id)
     {
+        _logger.LogInformation("Deleting employee with ID: {EmployeeId}", id);
+        
+        var employee = await _repository.GetByIdAsync(id);
+        if (employee == null)
+        {
+            _logger.LogWarning("Employee with ID {EmployeeId} not found", id);
+            return null;
+        }
+
         await _repository.DeleteAsync(id);
+        
+        _logger.LogInformation("Employee deleted successfully with ID: {EmployeeId}", id);
+        
+        return new EmployeeResponse
+        {
+            Id = employee.Id,
+            FirstName = employee.FirstName,
+            LastName = employee.LastName,
+            Position = employee.Position,
+            Email = employee.Email
+        };
     }
 }
